+++ src/components/errors/Error.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react'
-import classJoin from '../../utils/classJoin'
+import classNames from '../../utils/classNames'
 import './errors.css'
 
-export const Container = React.memo(function ({ className, children, ...props }: React.HTMLProps<HTMLDivElement>) {
-  return <div {...props} className={classJoin(['error-container', className])}>
+export const ErrorContainer = React.memo(function ({ className, children, ...props }: React.HTMLProps<HTMLDivElement>) {
+  return <div {...props} className={classNames(['error-container', className])}>
     <div className="error-background" />
     <div className="error-content">{children}</div>
   </div>
@@ -15,8 +15,8 @@ export type DetailsProps = React.HTMLProps<HTMLDivElement> & {
   description?: React.ReactNode,
 }
 
-export const Details = React.memo(function ({ className, header, description, backgroundHeader, children, ...props }: DetailsProps) {
-  return <div {...props} className={classJoin(['error-details', className])}>
+export const ErrorDetails = React.memo(function ({ className, header, description, backgroundHeader, children, ...props }: DetailsProps) {
+  return <div {...props} className={classNames(['error-details', className])}>
     {backgroundHeader && <div className="error-background-title">{backgroundHeader}</div>}
     <div className="error-title">{header || 'Something went wrong'}</div>
     {description && <div className="error-description">{description}</div>}
@@ -24,12 +24,8 @@ export const Details = React.memo(function ({ className, header, description, ba
   </div>
 })
 
-export const Image = React.memo(function ({ className, ...props }: React.ImgHTMLAttributes<HTMLImageElement>) {
-  return <div className={classJoin(['error-image', className])}>
+export const ErrorImage = React.memo(function ({ className, ...props }: React.ImgHTMLAttributes<HTMLImageElement>) {
+  return <div className={classNames(['error-image', className])}>
     <img alt="error" {...props} />
   </div>
 })
-
-export function reload() {
-  window.location.reload()
-}
