+++ src/components/errors/ErrorMetamaskLocked.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Container, Details, Image } from './Error'
+import { ErrorContainer, ErrorDetails, ErrorImage } from './Error'
 import { Button } from 'decentraland-ui/dist/components/Button/Button'
 
 import errorImage from '../../images/errors/robotsmiling.png'
@@ -11,12 +11,12 @@ export interface ErrorNetworkMismatchProps {
 }
 
 export const ErrorMetamaskLocked = React.memo(function (props: ErrorNetworkMismatchProps) {
-  return <Container id="error-metamask-locked">
-    <Details
+  return <ErrorContainer id="error-metamask-locked">
+    <ErrorDetails
       header="Before we continue"
       description={props.details}>
         <Button primary onClick={props.closeError}>Retry</Button>
-      </Details>
-    <Image alt="error-smiling-robot" src={errorImage} />
-  </Container>
+      </ErrorDetails>
+    <ErrorImage alt="error-smiling-robot" src={errorImage} />
+  </ErrorContainer>
 })
