+++ src/components/errors/ErrorNewLogin.tsx
@@ -1,13 +1,14 @@
 import React from 'react'
 import { Button } from 'decentraland-ui/dist/components/Button/Button'
-import { Container, Details, Image, reload } from './Error'
+import { ErrorContainer, ErrorDetails, ErrorImage } from './Error'
+import { reload } from "./utils"
 
 import errorImage from '../../images/errors/robotsmiling.png'
 import './errors.css'
 
 export const ErrorNewLogin = React.memo(function () {
-  return <Container id="error-new-login">
-    <Details
+  return <ErrorContainer id="error-new-login">
+    <ErrorDetails
       backgroundHeader="Oops!"
       header="Another session was detected"
       description={<>
@@ -15,7 +16,7 @@ export const ErrorNewLogin = React.memo(function () {
         <br />
         Please, close the prior session and click "Reload" to explore the world in this window.</>}>
       <Button primary onClick={reload}>Reload</Button>
-    </Details>
-    <Image alt="error-smiling-robot" src={errorImage} />
-  </Container>
+    </ErrorDetails>
+    <ErrorImage alt="error-smiling-robot" src={errorImage} />
+  </ErrorContainer>
 })
