+++ src/components/errors/ErrorNoMobile.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
-import { Container } from './Error'
+import { ErrorContainer } from './Error'
 import { Button } from 'decentraland-ui/dist/components/Button/Button'
 import errorImage from '../../images/errors/error-robotmobile.png'
 
 
 export const ErrorNoMobile = React.memo(function () {
-  return <Container id="error-no-mobile">
+  return <ErrorContainer id="error-no-mobile">
         <div className="error-no-mobile-title">The client is only available on desktop right now.</div>
         <div className="error-no-mobile-image">
           <img alt="no-mobile" width="495" height="707" src={errorImage} />
@@ -13,5 +13,5 @@ export const ErrorNoMobile = React.memo(function () {
         <div className="error-no-mobile-cta">
           <Button primary as="a" href="https://decentraland.org">Learn more about decentraland</Button>
         </div>
-  </Container>
+  </ErrorContainer>
 })
