+++ src/components/errors/ErrorNotSupported.tsx
@@ -1,15 +1,15 @@
 import React from 'react'
 import errorImage from '../../images/errors/error-robotmobile.png'
-import { Container, Details, Image } from './Error'
+import { ErrorContainer, ErrorDetails, ErrorImage } from './Error'
 import './errors.css'
 
 export const ErrorNotSupported = React.memo(function () {
-  return <Container id="error-new-login">
-    <Details
+  return <ErrorContainer id="error-new-login">
+    <ErrorDetails
       backgroundHeader="Error"
       header="Your browser or device is not supported"
       description="The Explorer only works on Chrome or Firefox for Windows, Linux and macOS."
       />
-    <Image alt="error-mobile-robot" src={errorImage} />
-  </Container>
+    <ErrorImage alt="error-mobile-robot" src={errorImage} />
+  </ErrorContainer>
 })
