+++ src/components/errors/utils.tsx
@@ -1,0 +1,4 @@
+
+export function reload() {
+  window.location.reload();
+}
